HMS improvements: - Find hospital by bed-type - See all shortage reports in one list + Some bug fixes in some modules.

+++ models/01_RESTlike_controller.py
@@ -1651,7 +1651,7 @@ def shn_update(jr, pheader=None, deletable=True, onvalidation=None, onaccept=Non
 
     else:
         record_id = jr.id
-        deletable = shn_has_permission('delete', table, record_id)
+        deletable = deletable and shn_has_permission('delete', table, record_id)
 
     authorised = shn_has_permission('update', table, record_id)
     if authorised:
@@ -1696,16 +1696,18 @@ def shn_update(jr, pheader=None, deletable=True, onvalidation=None, onaccept=Non
                 label_list_button = s3.crud_strings.label_list_button
             list_btn = A(label_list_button, _href=jr.there(), _id='list-btn')
 
-            del_href = jr.other(method='delete', representation=jr.representation)
-            if tablename in s3.crud_strings and "label_delete_button" in s3.crud_strings[tablename]:
-                label_del_button = s3.crud_strings[tablename].label_delete_button
-            else:
-                label_del_button = None
-            if label_del_button is None:
-                label_del_button = s3.crud_strings.label_delete_button
-            del_btn = A(label_del_button, _href=del_href, _id='delete-btn')
+            if deletable:
+                del_href = jr.other(method='delete', representation=jr.representation)
+                if tablename in s3.crud_strings and "label_delete_button" in s3.crud_strings[tablename]:
+                    label_del_button = s3.crud_strings[tablename].label_delete_button
+                else:
+                    label_del_button = None
+                if label_del_button is None:
+                    label_del_button = s3.crud_strings.label_delete_button
+                del_btn = A(label_del_button, _href=del_href, _id='delete-btn')
+                output.update(del_btn=del_btn)
 
-            output.update(title=title, list_btn=list_btn, del_btn=del_btn)
+            output.update(title=title, list_btn=list_btn)
 
             if jr.component:
                 # Block join field
